requirements.txt
Django>=5.0,<5.1
channels>=4.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
requests>=2.31.0
whitenoise>=6.6.0
python-dotenv>=1.0.0
uvicorn>=0.27.0 